Renamed workout screen folder to fit naming convention
